v2.1 ready to go, updated all comments

+++ contracts/interface/IMAX721.sol
@@ -22,9 +22,10 @@ pragma solidity >=0.8.0 <0.9.0;
 
 import "@openzeppelin/contracts/utils/introspection/IERC165.sol";
 
-  ///
-  /// Developer this is the standard interface for all ERC721's written by myself
-  ///
+///
+/// @dev Interface for @MaxFlowO2's Contracts
+///  must include or add totalSupply() to main
+///
 
 interface IMAX721 is IERC165 {
 
@@ -39,30 +40,37 @@ interface IMAX721 is IERC165 {
   // IMAX721 => 0x29499a25
 
   // @notice will return status of Minter
-  // minterStatus() => 0x2ecd28ab
+  // @return - bool of active or not
+  // ERC165 datum minterStatus() => 0x2ecd28ab
   function minterStatus() external view returns (bool);
 
   // @notice will return minting fees
-  // minterFees() => 0xd95ae162
-  function minterFees() external view returns (uint256);
+  // @return - uint of mint costs in wei
+  // ERC165 datum minterFees() => 0xd95ae162
+  function minterFees() external view returns (uint);
 
   // @notice will return maximum mint capacity
-  // minterMaximumCapacity() => 0x78c5939b
-  function minterMaximumCapacity() external view returns (uint256);
+  // @return - uint of maximum mints allowed
+  // ERC165 datum minterMaximumCapacity() => 0x78c5939b
+  function minterMaximumCapacity() external view returns (uint);
 
   // @notice will return maximum "team minting" capacity
-  // minterMaximumTeamMints() => 0x049157bb
-  function minterMaximumTeamMints() external view returns (uint256);
+  // @return - uint of maximum airdrops or team mints allowed
+  // ERC165 datum minterMaximumTeamMints() => 0x049157bb
+  function minterMaximumTeamMints() external view returns (uint);
 
   // @notice will return "team mints" left
-  // minterTeamMintsRemaining() => 0x5c17e370
-  function minterTeamMintsRemaining() external view returns (uint256);
+  // @return - uint of remaing airdrops or team mints
+  // ERC165 datum minterTeamMintsRemaining() => 0x5c17e370
+  function minterTeamMintsRemaining() external view returns (uint);
 
   // @notice will return "team mints" count
-  // minterTeamMintsCount() => 0xe68b7961
-  function minterTeamMintsCount() external view returns (uint256);
+  // @return - uint of airdrops or team mints done
+  // ERC165 datum minterTeamMintsCount() => 0xe68b7961
+  function minterTeamMintsCount() external view returns (uint);
 
   // @notice will return current token count
-  // totalSupply() => 0x18160ddd
-  function totalSupply() external view returns (uint256);
+  // @return - uint of how many NFT's minted on contract
+  // ERC165 datum totalSupply() => 0x18160ddd
+  function totalSupply() external view returns (uint);
 }
